+++ iceoryx_dds/test/mocks/google_mocks.hpp
@@ -51,8 +51,8 @@ class MockDataReader
 public:
     MockDataReader(const iox::capro::ServiceDescription& sd){};
     MOCK_METHOD0(connect, void(void));
-    MOCK_METHOD3(read, iox::cxx::expected<uint8_t, iox::dds::DataReaderError>(uint8_t* const buffer, const uint64_t&, const uint64_t&));
-    MOCK_METHOD4(read, iox::cxx::expected<uint8_t, iox::dds::DataReaderError>(uint8_t* const buffer, const uint64_t&, const uint64_t&, const uint64_t&));
+    MOCK_METHOD3(read, iox::cxx::expected<uint64_t, iox::dds::DataReaderError>(uint8_t* const buffer, const uint64_t&, const uint64_t&));
+    MOCK_METHOD4(read, iox::cxx::expected<uint64_t, iox::dds::DataReaderError>(uint8_t* const buffer, const uint64_t&, const uint64_t&, const uint64_t&));
     MOCK_CONST_METHOD0(getServiceId, std::string(void));
     MOCK_CONST_METHOD0(getInstanceId, std::string(void));
     MOCK_CONST_METHOD0(getEventId, std::string(void));
